+++ src/pages/SelectPage.js
@@ -1,5 +1,7 @@
-import React, { useState } from 'react';
-import { Select, Table, Card } from '../components';
+import React, { useState, createRef } from 'react';
+import { Select, Table, Card, Collapse, Icon, BackTopBtn, Tag } from '../components';
+import { Prism as SyntaxHighlighter } from 'react-syntax-highlighter';
+import { prism } from 'react-syntax-highlighter/dist/esm/styles/prism';
 
 const keys = ['property', 'description', 'default', 'type', 'value'];
 const rows = [
@@ -11,20 +13,6 @@ const rows = [
         value: ''
     },
     { 
-        property: 'selectedItem', 
-        description: 'Selected option', 
-        default: '', 
-        type: 'string | number',
-        value: ''
-    },
-    { 
-        property: 'onSelect', 
-        description: 'Handle selected value', 
-        default: '', 
-        type: 'function',
-        value: ''
-    },
-    { 
         property: 'itemName', 
         description: 'Set if you pass items as array of objects', 
         default: '', 
@@ -32,11 +20,46 @@ const rows = [
         value: ''
     },
     { 
-        property: 'addSubItem', 
-        description: 'Set second additional prop of object if you pass items as array of objects', 
+        property: 'onChange', 
+        description: 'Invokes in select option select (return item value)', 
+        default: '', 
+        type: 'function',
+        value: ''
+    },
+    { 
+        property: 'childrenKey', 
+        description: 'Set if you pass items with children array (pass the key of item that has children array)', 
         default: '', 
         type: 'string',
         value: ''
+    },
+    { 
+        property: 'searchable', 
+        description: 'Enable searching field in select menu (has effect only with one level list)', 
+        default: 'false', 
+        type: 'boolean',
+        value: 'true | false'
+    },
+    { 
+        property: 'search', 
+        description: 'Pass search state value', 
+        default: '', 
+        type: 'string',
+        value: ''
+    },
+    { 
+        property: 'onSearch', 
+        description: 'Invokes on search value change (return search value)', 
+        default: '', 
+        type: 'function',
+        value: ''
+    },
+    { 
+        property: 'size', 
+        description: 'Set select size', 
+        default: '', 
+        type: 'string',
+        value: 'medium | large'
     },
     { 
         property: 'width', 
@@ -46,14 +69,7 @@ const rows = [
         value: '100px | 100% | 100'
     },
     { 
-        property: 'minWidth', 
-        description: 'Set select min width', 
-        default: 'auto', 
-        type: 'string | number',
-        value: '100px | 100% | 100'
-    },
-    { 
-        property: 'labelName', 
+        property: 'label', 
         description: 'Set select label', 
         default: '', 
         type: 'string',
@@ -68,7 +84,7 @@ const rows = [
     },
     { 
         property: 'color', 
-        description: 'Set color of pagination from list', 
+        description: 'Set select color', 
         default: '', 
         type: 'string',
         value: 'primary | info | success | error'
@@ -81,18 +97,11 @@ const rows = [
         value: 'true | false'
     },
     { 
-        property: 'subItem', 
-        description: 'Set item name in subitems menu',
-        default: 'false', 
-        type: 'boolean',
-        value: 'true | false'
-    },
-    { 
-        property: 'size', 
-        description: 'Set select size',
+        property: 'prefix', 
+        description: 'Set prefix',
         default: '', 
-        type: 'string',
-        value: 'medium | large'
+        type: 'any',
+        value: ''
     },
     { 
         property: 'className',
@@ -103,7 +112,7 @@ const rows = [
     }
 ]
 
-const items = [
+const countries = [
     { country: 'Kazakhstan', cities: [ 'Nur-Sultan', 'Almaty', 'Shymkent' ] }, 
     { country: 'Russia', cities: ['Moscow', 'St. Petersburg', 'Krasnodar'] }, 
     { country: 'USA', cities: ['Washington, D.C.', 'New York City', 'San Francisco'] }, 
@@ -111,124 +120,199 @@ const items = [
     { country: 'China', cities: ['Beijing', 'Shanghai', 'Shenzhen'] }
 ]
 
-export const SelectPage = () => {
-    const [selected, setSelected] = useState('');
-    const [selectedCascader, setSelectedCascader] = useState('');
-    const [selectedCascaderSubItem, setSelectedCascaderSubItem] = useState('');
+const simpleUsage = 
+`// Usage examples
+import React, { useState } from 'react';
+import { Select } from '@assenti/react-ui-components';
+
+const countries = [
+    { country: 'Kazakhstan', cities: [ 'Nur-Sultan', 'Almaty', 'Shymkent' ] }, 
+    { country: 'Russia', cities: ['Moscow', 'St. Petersburg', 'Krasnodar'] }, 
+    { country: 'USA', cities: ['Washington, D.C.', 'New York City', 'San Francisco'] }, 
+    { country: 'United Kingdom', cities: ['London', 'York', 'Manchester'] }, 
+    { country: 'China', cities: ['Beijing', 'Shanghai', 'Shenzhen'] }
+]
+
+function Example() {
+    const [city, setCity] = useState('');
+    const [country, setCountry] = useState('');
 
     return (
-        <div className="rui-page">
-            <div className="rui-page-title">Selects</div>
+        <div className="pa-20">
+            <Select
+                items={countries}
+                prefix={<Icon name="earth"/>}
+                itemTitle="country"
+                label="Select your favourite country"
+                width={250}
+                color="info"
+                placeholder="Countries"
+                value={country}
+                onChange={v => setCountry(v)}/>
+            <Select
+                className="ml-20"
+                label="Select your favourite city"
+                items={countries}
+                itemTitle="country"
+                childrenKey="cities"
+                color="primary"
+                width={250}
+                placeholder="Cities"
+                value={city}
+                onChange={v => setCity(v)}/>
+        </div>
+    )
+}`
+
+const searchUsage = 
+`// Usage examples
+import React, { useState } from 'react';
+import { Select } from '@assenti/react-ui-components';
+
+const countries = [
+    { country: 'Kazakhstan', cities: [ 'Nur-Sultan', 'Almaty', 'Shymkent' ] }, 
+    { country: 'Russia', cities: ['Moscow', 'St. Petersburg', 'Krasnodar'] }, 
+    { country: 'USA', cities: ['Washington, D.C.', 'New York City', 'San Francisco'] }, 
+    { country: 'United Kingdom', cities: ['London', 'York', 'Manchester'] }, 
+    { country: 'China', cities: ['Beijing', 'Shanghai', 'Shenzhen'] }
+]
+
+function Example() {
+    const [country_, setCountry_] = useState('');
+    const [search, setSearch] = useState('');
+
+    return (
+        <div className="pa-20">
+            <Select
+                items={countries}
+                prefix={<Icon name="earth"/>}
+                itemTitle="country"
+                label="Select your favourite country"
+                width={250}
+                search={search}
+                searchable
+                onSearch={value => setSearch(value)}
+                color="info"
+                placeholder="Countries"
+                value={country_}
+                onChange={value => setCountry_(value)}/>
+            <Select
+                items={countries}
+                prefix={<Icon name="earth"/>}
+                itemTitle="country"
+                label="Select your favourite country"
+                width={250}
+                search={search}
+                searchable
+                rounded
+                className="ml-20"
+                onSearch={value => setSearch(value)}
+                color="info"
+                placeholder="Countries"
+                value={country_}
+                onChange={value => setCountry_(value)}/>
+        </div>
+    )
+}`
+
+export const SelectPage = () => {
+    const [city, setCity] = useState('');
+    const [country, setCountry] = useState('');
+    const [country_, setCountry_] = useState('');
+    const [search, setSearch] = useState('');
+    const parent = createRef();
+    const api = createRef();
+
+    const goToApi = () => {
+        if (api.current) api.current.scrollIntoView({ behavior: 'smooth', block: 'center' })
+    }
+
+    return (
+        <div className="rui-page" ref={parent}>
+            <div className="row align-center space-between">
+                <div className="rui-page-title">
+                    Select Component <Tag iconLeft="hammer" value="WIP" small color="error"/>
+                </div>
+                <div className="rui-link fz-13 fw-bold mr-10" onClick={() => goToApi()}>API</div>
+            </div>
             <Card outlined color="primary" title="Simple Selects">
                 <Select
-                    items={items}
+                    items={countries}
+                    prefix={<Icon name="earth"/>}
+                    itemTitle="country"
+                    label="Select your favourite country"
+                    width={250}
+                    color="info"
+                    placeholder="Countries"
+                    value={country}
+                    onChange={v => setCountry(v)}/>
+                <Select
+                    className="ml-20"
+                    label="Select your favourite city"
+                    items={countries}
+                    itemTitle="country"
+                    childrenKey="cities"
                     color="primary"
-                    width={200}
-                    itemName="country"
-                    placeholder="Countries"
-                    selectedItem={selected}
-                    onSelect={v => setSelected(v)}/>
-                <Select
-                    items={items}
-                    labelName="Rounded select"
-                    width={200}
-                    color="info"
-                    rounded
-                    itemName="country"
-                    placeholder="Countries"
-                    selectedItem={selected}
-                    onSelect={v => setSelected(v)}/>
-                <Select
-                    items={items}
-                    width={200}
-                    color="success"
-                    itemName="country"
-                    placeholder="Countries"
-                    selectedItem={selected}
-                    onSelect={v => setSelected(v)}/>
-                <Select
-                    items={items}
-                    width={200}
-                    color="error"
-                    itemName="country"
-                    placeholder="Countries"
-                    selectedItem={selected}
-                    onSelect={v => setSelected(v)}/>
-                <Select
-                    placeholder="Select a country" 
-                    items={items}
-                    width={200}
-                    size="medium"
-                    itemName="country"
-                    labelName="Medium size select"
-                    selectedItem={selected}
-                    onSelect={v => setSelected(v)}/>
-                <Select
-                    placeholder="Select a country" 
-                    items={items}
-                    width={200}
-                    size="medium"
-                    rounded
-                    color="info"
-                    itemName="country"
-                    labelName="Rounded medium size select"
-                    selectedItem={selected}
-                    onSelect={v => setSelected(v)}/>
-                <Select
-                    placeholder="Select a country" 
-                    items={items}
-                    width={200}
-                    size="large"
-                    itemName="country"
-                    labelName="Large size select"
-                    selectedItem={selected}
-                    onSelect={v => setSelected(v)}/>
+                    width={250}
+                    placeholder="Cities"
+                    value={city}
+                    onChange={v => setCity(v)}/>
+                <Collapse 
+                    icon="code" 
+                    iconSize={18}
+                    contentStyles={{ padding: 0 }}
+                    tooltip="Code">
+                    <SyntaxHighlighter language="jsx" style={prism}>
+                        {simpleUsage}
+                    </SyntaxHighlighter> 
+                </Collapse>
             </Card>
             <br/>
-            <Card outlined color="primary" title="Selects with subitems">
+            <Card outlined color="primary" title="Simple Select with searching">
                 <Select
-                    items={items}
-                    width={200}
-                    selectedItem={selectedCascader}
-                    placeholder="Select a country and a capital" 
-                    itemName="country"
-                    subItemsField="cities"
+                    items={countries}
+                    prefix={<Icon name="earth"/>}
+                    itemTitle="country"
+                    label="Select your favourite country"
+                    width={250}
+                    size="medium"
+                    search={search}
+                    searchable
+                    onSearch={value => setSearch(value)}
                     color="info"
-                    labelName="Countries and capitals"
-                    onSelect={v => setSelectedCascader(v)}/>
-            </Card>
-            <br/>
-            <Card outlined color="primary" title="Selects with subitems (show only subitem value)">
+                    placeholder="Countries"
+                    value={country_}
+                    onChange={value => setCountry_(value)}/>
                 <Select
-                    items={items}
-                    width={200}
-                    selectedItem={selectedCascaderSubItem}
-                    placeholder="Select a country and a capital" 
-                    itemName="country"
-                    subItemsField="cities"
-                    subTitle
-                    showSubItemOnly={true}
-                    labelName="Countries and capitals"
-                    color="info"
-                    onSelect={v => setSelectedCascaderSubItem(v)}/>
-                <Select
-                    items={items}
-                    width={200}
-                    selectedItem={selectedCascaderSubItem}
-                    placeholder="Select a country and a capital" 
-                    itemName="country"
-                    subItemsField="cities"
-                    subTitle
+                    items={countries}
+                    prefix={<Icon name="earth"/>}
+                    itemTitle="country"
+                    label="Select your favourite country"
+                    width={250}
+                    search={search}
+                    searchable
                     rounded
-                    showSubItemOnly={true}
-                    labelName="Countries and capitals"
+                    className="ml-20"
+                    onSearch={value => setSearch(value)}
                     color="info"
-                    onSelect={v => setSelectedCascaderSubItem(v)}/>
+                    placeholder="Countries"
+                    value={country_}
+                    onChange={value => setCountry_(value)}/>
+                <Collapse 
+                    icon="code" 
+                    iconSize={18}
+                    contentStyles={{ padding: 0 }}
+                    tooltip="Code">
+                    <SyntaxHighlighter language="jsx" style={prism}>
+                        {searchUsage}
+                    </SyntaxHighlighter> 
+                </Collapse>
             </Card>
-            <h2>API</h2>
+            <BackTopBtn setRef={parent} dark size="medium"/>
+            <h2 ref={api}>API</h2>
             <Table
                 bordered
-                headers={keys}
+                headers={['Property', 'Description', 'Default', 'Type', 'Value']}
                 items={rows}
                 index={true}
                 itemTitles={keys}/>